# arrow: use dictionary array for reference text column