@HashidaTKS 2021-09-30
apache/arrow#10990 is merged.
Please merge this after next version Apache Arrow is released if there is no problem.

I am checking the failed checks, it seems like environment problem...