@kou 2021-09-30
Could you rebase on master after Apache Arrow 6.0.0 is released?